refactor(basis): move private _basis_*.py modules into basis/ package
Converts basis.py to basis/__init__.py and moves the 5 private
_basis_*.py implementation files into the package, mirroring the
bspline/ and bezier/ package structure.
