+++ src/pantr/quad.py
@@ -342,7 +342,7 @@ def create_lagrange_points_lattice(
         ValueError: If any value in ``n_pts_per_dir`` is less than 1.
     """
     # Lazy import to avoid circular dependency
-    from ._basis_lagrange import _get_lagrange_points  # noqa: PLC0415
+    from .basis._basis_lagrange import _get_lagrange_points  # noqa: PLC0415
 
     if any(n_pts < 1 for n_pts in n_pts_per_dir):
         raise ValueError("All number of points must be at least 1")
